Added duplicate field support and fixed issue with child name same as parent name.
Fixed issue where fields with the same name as the split field would not
be present in the final output events (Example: “location”:
{“location”:”New York”,”lat”:38.1234,”lon”:-121.1234} )
Changed the childindex field name to childindex__# where the # is now
unique.
If multiple field splitters are used in a row, each one will create a
new childindex__# field.
Avoids the “duplicate field” error previously encountered or the
requirement to field map between splitters to rename the childindex
field.
Duplicate field names from multi-cardinal  groups will be renamed with a
unique __# appended to the end of the field name.